ci: split api_2 test
