+++ packages/amplify-e2e-tests/src/__tests__/api_2a.test.ts
@@ -6,14 +6,11 @@ import {
   createNewProjectDir,
   deleteProject,
   deleteProjectDir,
-  enableAdminUI,
   getAppSyncApi,
-  getLocalEnvInfo,
   getProjectMeta,
   getTransformConfig,
   initJSProjectWithProfile,
   updateApiSchema,
-  updateAPIWithResolutionStrategyWithModels,
 } from '@aws-amplify/amplify-e2e-core';
 import AWSAppSyncClient, {
   AUTH_TYPE,
@@ -173,78 +170,4 @@ describe('amplify add api (GraphQL)', () => {
     expect(_.isEmpty(disableDSConfig.ResolverConfig)).toBe(true);
   });
 
-  it('init a project with conflict detection enabled and admin UI enabled to generate DataStore models in the cloud', async () => {
-    // eslint-disable-next-line spellcheck/spell-checker
-    const name = 'dsadminui';
-    await initJSProjectWithProfile(projRoot, { disableAmplifyAppCreation: false, name });
-
-    let meta = getProjectMeta(projRoot);
-    const appId = meta.providers?.[providerName]?.AmplifyAppId;
-    const region = meta.providers?.[providerName]?.Region;
-
-    expect(appId).toBeDefined();
-
-    const localEnvInfo = getLocalEnvInfo(projRoot);
-    const { envName } = localEnvInfo;
-
-    // setup Amplify Studio backend
-    await enableAdminUI(appId, envName, region);
-
-    await addApiWithBlankSchemaAndConflictDetection(projRoot, { transformerVersion: 1 });
-    await updateApiSchema(projRoot, name, 'simple_model.graphql');
-    await amplifyPush(projRoot);
-
-    meta = getProjectMeta(projRoot);
-
-    const { output } = meta.api[name];
-    const { GraphQLAPIIdOutput, GraphQLAPIEndpointOutput, GraphQLAPIKeyOutput } = output;
-
-    expect(GraphQLAPIIdOutput).toBeDefined();
-    expect(GraphQLAPIEndpointOutput).toBeDefined();
-    expect(GraphQLAPIKeyOutput).toBeDefined();
-
-    const { graphqlApi } = await getAppSyncApi(GraphQLAPIIdOutput, meta.providers.awscloudformation.Region);
-
-    expect(graphqlApi).toBeDefined();
-    expect(graphqlApi.apiId).toEqual(GraphQLAPIIdOutput);
-  });
-
-  it('init a sync enabled project and update conflict resolution strategy', async () => {
-    // eslint-disable-next-line spellcheck/spell-checker
-    const name = 'syncenabled';
-    await initJSProjectWithProfile(projRoot, { name });
-    await addApiWithBlankSchemaAndConflictDetection(projRoot, { transformerVersion: 1 });
-    await updateApiSchema(projRoot, name, 'simple_model.graphql');
-
-    let transformConfig = getTransformConfig(projRoot, name);
-    expect(transformConfig).toBeDefined();
-    expect(transformConfig.ResolverConfig).toBeDefined();
-    expect(transformConfig.ResolverConfig.project).toBeDefined();
-    expect(transformConfig.ResolverConfig.project.ConflictDetection).toEqual('VERSION');
-    expect(transformConfig.ResolverConfig.project.ConflictHandler).toEqual('AUTOMERGE');
-
-    await updateAPIWithResolutionStrategyWithModels(projRoot, {});
-
-    transformConfig = getTransformConfig(projRoot, name);
-    expect(transformConfig).toBeDefined();
-    expect(transformConfig.Version).toBeDefined();
-    expect(transformConfig.Version).toEqual(TRANSFORM_CURRENT_VERSION);
-    expect(transformConfig.ResolverConfig).toBeDefined();
-    expect(transformConfig.ResolverConfig.project).toBeDefined();
-    expect(transformConfig.ResolverConfig.project.ConflictDetection).toEqual('VERSION');
-    expect(transformConfig.ResolverConfig.project.ConflictHandler).toEqual('OPTIMISTIC_CONCURRENCY');
-
-    await amplifyPush(projRoot);
-    const meta = getProjectMeta(projRoot);
-    const { output } = meta.api[name];
-    const { GraphQLAPIIdOutput, GraphQLAPIEndpointOutput, GraphQLAPIKeyOutput } = output;
-    const { graphqlApi } = await getAppSyncApi(GraphQLAPIIdOutput, meta.providers.awscloudformation.Region);
-
-    expect(GraphQLAPIIdOutput).toBeDefined();
-    expect(GraphQLAPIEndpointOutput).toBeDefined();
-    expect(GraphQLAPIKeyOutput).toBeDefined();
-
-    expect(graphqlApi).toBeDefined();
-    expect(graphqlApi.apiId).toEqual(GraphQLAPIIdOutput);
-  });
 });
